Use localstorage for persistent storage
