+++ src/open_in_gitpod.tsx
@@ -30,7 +30,8 @@ import createWorksapceFromContext from "./helpers/createWorkspaceFromContext";
 import { useIssueHistory } from "./helpers/issue";
 import OpenInGitpod, { getPreferencesForContext } from "./helpers/openInGitpod";
 import { usePullReqHistory } from "./helpers/pull-request";
-import { useHistory, useFavorites } from "./helpers/repository";
+import { useHistory } from "./helpers/repository";
+import { useFavorites } from "./helpers/repository-favorites";
 import { getGitHubClient } from "./helpers/withGithubClient";
 import { dashboardPreferences } from "./preferences/dashboard_preferences";
 import RepositoryPreference from "./preferences/repository_preferences";
